Move FileContent and implementations to utils package
* Created new `utils` package.
* Moved `FileContent` interface, `WeakFileContent`, `LazyFileContent`, and `MemoryFileContent` to `utils/file_content.go`.
* Moved corresponding tests to `utils/file_content_test.go`.
* Removed these structs from `site.go`.
* Fixes PR feedback.
