+++ utils/file_content_test.go
@@ -1,4 +1,4 @@
-package g2
+package utils
 
 import (
 	"bytes"
